relocate articlecategory model from mainapp to article
