+++ article/models.py
@@ -1,7 +1,25 @@
 from django.db import models
-
-from mainapp.models import ArticleCategory
 from authapp.models import HabrUser
+
+
+class ArticleCategory(models.Model):
+    name = models.CharField(max_length=64, unique=True, verbose_name='название')
+    description = models.TextField(verbose_name='описание')
+    is_active = models.BooleanField(default=True)
+
+    def __str__(self):
+        return f'{self.name}'
+
+    class Meta:
+        verbose_name = 'категория'
+        verbose_name_plural = 'категории'
+        ordering = ('-id',)
+
+    def delete(self):
+        if self.is_active:
+            self.is_active = False
+
+        self.save()
 
 
 class Article(models.Model):
@@ -22,5 +40,5 @@ class Article(models.Model):
     def delete(self):
         if self.is_active:
             self.is_active = False
-
         self.save()
+
