+++ mainapp/models.py
@@ -1,21 +1,1 @@
 from django.db import models
-
-
-class ArticleCategory(models.Model):
-    name = models.CharField(max_length=64, unique=True, verbose_name='название')
-    description = models.TextField(verbose_name='описание')
-    is_active = models.BooleanField(default=True)
-
-    def __str__(self):
-        return f'{self.name}'
-
-    class Meta:
-        verbose_name = 'категория'
-        verbose_name_plural = 'категории'
-        ordering = ('-id',)
-
-    def delete(self):
-        if self.is_active:
-            self.is_active = False
-
-        self.save()
