+++ mainapp/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from django.shortcuts import get_object_or_404
 
-from mainapp.models import ArticleCategory
+from article.models import ArticleCategory
 from article.models import Article
